requirements.txt
pandas
numpy
mysql-connector-python
pymysql
scikit-learn
seaborn
matplotlib
lazypredict
python-dotenv
-e .